requirements.txt
langchain
openai
psycopg2-binary
pandas
matplotlib
python-dotenv
scikit-learn
faker
pydantic
uvicorn
fastapi
# Dodano bibliotekę do generowania wykresów
plotly
kaleido
